Bedrock 1.20.70

+++ src/main/java/cn/nukkit/block/BlockAir.java
@@ -59,12 +59,12 @@ public class BlockAir extends BlockTransparent {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return 0;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 0;
     }
