+++ src/main/java/cn/nukkit/block/BlockBorder.java
@@ -24,12 +24,12 @@ public class BlockBorder extends BlockWall {
     }
 
     @Override
-    public double getHardness() {
-        return 0.2;
+    public float getHardness() {
+        return 0.2f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 18000;
     }
