+++ src/main/java/cn/nukkit/block/BlockDaylightDetector.java
@@ -42,12 +42,12 @@ public class BlockDaylightDetector extends BlockTransparentMeta {
     }
 
     @Override
-    public double getHardness() {
-        return 0.2;
+    public float getHardness() {
+        return 0.2f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 1;
     }
