+++ src/main/java/cn/nukkit/block/BlockFroglight.java
@@ -6,13 +6,13 @@ public abstract class BlockFroglight extends BlockRotatedPillar {
     }
 
     @Override
-    public double getHardness() {
-        return 0.3;
+    public float getHardness() {
+        return 0.3f;
     }
 
     @Override
-    public double getResistance() {
-        return 1.5;
+    public float getResistance() {
+        return 1.5f;
     }
 
     @Override
