+++ src/main/java/cn/nukkit/block/BlockGlassTinted.java
@@ -17,13 +17,13 @@ public class BlockGlassTinted extends BlockTransparent {
     }
 
     @Override
-    public double getHardness() {
-        return 0.3;
+    public float getHardness() {
+        return 0.3f;
     }
 
     @Override
-    public double getResistance() {
-        return 1.5;
+    public float getResistance() {
+        return 1.5f;
     }
 
     @Override
