+++ src/main/java/cn/nukkit/block/BlockGrass.java
@@ -35,12 +35,12 @@ public class BlockGrass extends BlockDirt {
     }
 
     @Override
-    public double getHardness() {
-        return 0.6;
+    public float getHardness() {
+        return 0.6f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 3;
     }
