+++ src/main/java/cn/nukkit/block/BlockHoneycomb.java
@@ -18,12 +18,12 @@ public class BlockHoneycomb extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
-        return 0.6;
+    public float getHardness() {
+        return 0.6f;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 3;
     }
