+++ src/main/java/cn/nukkit/block/BlockJigsaw.java
@@ -43,12 +43,12 @@ public class BlockJigsaw extends BlockSolidMeta implements Faceable {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return -1;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 18000000;
     }
 
@@ -132,6 +132,7 @@ public class BlockJigsaw extends BlockSolidMeta implements Faceable {
             return true;
         }
         //TODO: UI
+        //player.openBlockEditor(getFloorX(), getFloorY(), getFloorZ(), ContainerType.JIGSAW_EDITOR);
         return true;
     }
