+++ src/main/java/cn/nukkit/block/BlockOreCoalDeepslate.java
@@ -17,8 +17,8 @@ public class BlockOreCoalDeepslate extends BlockOreCoal {
     }
 
     @Override
-    public double getHardness() {
-        return 4.5;
+    public float getHardness() {
+        return 4.5f;
     }
 
     @Override
