+++ src/main/java/cn/nukkit/block/BlockOreCopperDeepslate.java
@@ -17,8 +17,8 @@ public class BlockOreCopperDeepslate extends BlockOreCopper {
     }
 
     @Override
-    public double getHardness() {
-        return 4.5;
+    public float getHardness() {
+        return 4.5f;
     }
 
     @Override
