+++ src/main/java/cn/nukkit/block/BlockOreGoldDeepslate.java
@@ -17,8 +17,8 @@ public class BlockOreGoldDeepslate extends BlockOreGold {
     }
 
     @Override
-    public double getHardness() {
-        return 4.5;
+    public float getHardness() {
+        return 4.5f;
     }
 
     @Override
