+++ src/main/java/cn/nukkit/block/BlockOreLapisDeepslate.java
@@ -17,8 +17,8 @@ public class BlockOreLapisDeepslate extends BlockOreLapis {
     }
 
     @Override
-    public double getHardness() {
-        return 4.5;
+    public float getHardness() {
+        return 4.5f;
     }
 
     @Override
