+++ src/main/java/cn/nukkit/block/BlockReserved6.java
@@ -16,12 +16,12 @@ public class BlockReserved6 extends BlockSolid {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return 0;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 0;
     }
 }
