+++ src/main/java/cn/nukkit/block/BlockStructure.java
@@ -45,12 +45,12 @@ public class BlockStructure extends BlockSolidMeta {
     }
 
     @Override
-    public double getHardness() {
+    public float getHardness() {
         return -1;
     }
 
     @Override
-    public double getResistance() {
+    public float getResistance() {
         return 18000000;
     }
 
@@ -118,6 +118,7 @@ public class BlockStructure extends BlockSolidMeta {
             return true;
         }
         //TODO: UI
+        //player.openBlockEditor(getFloorX(), getFloorY(), getFloorZ(), ContainerType.STRUCTURE_EDITOR);
         return true;
     }
